Fix costbug

+++ express-back-end/database/schema/01_items.sql
@@ -3,7 +3,7 @@ CREATE TABLE items (
   id SERIAL PRIMARY KEY NOT NULL,
   title VARCHAR(255) NOT NULL,
   description VARCHAR(255) NOT NULL,
-  cost INT,
+  cost FLOAT,
   is_deleted BOOLEAN DEFAULT 'false',
   comment VARCHAR(255)
 );
